requirements.txt
flask
psutil
py-cpuinfo
matplotlib
reportlab
bcrypt